server/requirements.txt
certifi==2019.3.9
Django==2.2.1
djangorestframework==3.9.4
protobuf==3.7.1
psycopg2==2.8.2
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
uWSGI==2.0.18
